Scripts to turn GTFS static info into feed/sequences data
This is part of the foundation of adding support for multiple
feeds. Since the UI will still involve picking a station, we need to
know what feeds that station might appear in. We also want to organize
the station picker by route.

So we add this data file, which groups a route ID with a feed and a
list of stop IDs. When the user picks a stop ID, we'll have to fetch
all the feeds in which it might appear.